refactor: reorganize project layers
