+++ src/adapters/api/mod.rs
