+++ src/adapters/mod.rs
@@ -1,1 +1,2 @@
 pub mod repositories;
+pub mod api;
