+++ src/domain/mod.rs
@@ -1,3 +1,2 @@
 pub mod models;
 pub mod services;
-pub mod interfaces;
